Extract acquire into a standalone module for reuse outside acceptance tests
- moves the acquire package from acceptance/internal/acquire into a new
  github.com/openziti/sdk-golang/acquire module with its own go.mod/go.sum
- updates the acceptance harness and matrix command to import the new path
- adds a require plus in-repo replace for the acquire module to acceptance
- relocates the checked-in versions.yaml validation test into the acceptance
  harness package, locating the file via FindVersionsFile instead of a fixed path
- refreshes the package doc to describe the general-purpose, caller-parameterized surface
