+++ acquire/acquire.go
