+++ acquire/acquire_test.go
