+++ acquire/cache.go
