+++ acquire/extract.go
