+++ acquire/go.mod
@@ -1,0 +1,14 @@
+module github.com/openziti/sdk-golang/acquire
+
+go 1.25.0
+
+require (
+	github.com/stretchr/testify v1.11.1
+	golang.org/x/mod v0.37.0
+	gopkg.in/yaml.v3 v3.0.1
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+)
