+++ acquire/releases.go
