+++ acquire/releases_test.go
