+++ acquire/resolve.go
