+++ acquire/resolve_test.go
