+++ acquire/spec.go
