+++ acquire/spec_test.go
